Create AlgortihmHw
The algorithm checks the equality of two finite automaton. 
Program will get two file names as parameters. The file structure that define the DFA shown in
figure 1 is as follows (a=0,b=1,c=2,d=3 for the first file):
The first DFA:
4 2 0 2
1 3
2 3
3 3
3 3
The second DFA:
4 2 0 2
1 3
3 2
3 3
3 3
First line of the file gives the total number of states. Second line is the number of symbols. Third
line gives the starting state. Fourth line gives the final states, which can be more than one. Rest of the
lines gives the arcs for each of the states in order. Each number in a line gives the arc for one symbol.
For example for the first DFA, fifth line is ’1 3’, which means state 0 goes to 1 for the first symbol and
goes to 3 for the second symbol. States are numbered starting from 0.